refactor: extract shared auth library from solr-search (#1288)
Create src/aithena-common/ package with password hashing (passwords.py)
and auth DB schema management (auth_db.py) extracted from
src/solr-search/auth.py.

- installer/setup.py: remove sys.path hack and load_auth_helpers(),
  import directly from aithena_common
- installer/pyproject.toml: new uv project depending on aithena-common
- src/solr-search/auth.py: delegate to aithena_common for hash_password,
  verify_password, init_auth_db schema, SCHEMA_VERSION, get_schema_version
- src/solr-search/pyproject.toml: add aithena-common dependency
- ruff.toml: add aithena_common to known-first-party
- installer/tests/conftest.py: update mocks for new import structure

All 1026 solr-search tests and 6 installer tests pass.

Closes #1288
