+++ ruff.toml
@@ -7,6 +7,7 @@ select = ["E", "F", "W", "I", "UP", "B", "SIM", "S"]
 [lint.isort]
 known-first-party = [
     "admin_auth",
+    "aithena_common",
     "auth",
     "auth_helpers",
     "backup_service",
