+++ src/aithena-common/aithena_common/__init__.py
@@ -1,0 +1,1 @@
+"""Shared utilities for Aithena services."""
