+++ src/solr-search/pyproject.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 description = "Solr-backed FastAPI search service for aithena"
 requires-python = ">=3.12"
 dependencies = [
+    "aithena-common",
     "fastapi[all]>=0.115.0,<1",
     "uvicorn[standard]>=0.30.0,<1",
     "requests==2.33.0",
@@ -14,6 +15,9 @@ dependencies = [
     "pyjwt>=2.8.0",
     "python-json-logger>=4.0.0",
 ]
+
+[tool.uv.sources]
+aithena-common = { path = "../aithena-common", editable = true }
 
 [dependency-groups]
 dev = [
